Fix compatibility between account_move_chatter and invoice_refund_not_earlier
The invoice_refund_not_earlier module should only raise a constraint when a journal entry is reversed using the reversal wizard.
It this case, the method reverse_moves is called with a date.
If the method reverse_moves is called without parameters, the current date is used.
This behavior should not be impacted.