14.0 develop (#75)
* TA#35238 Port modules to v14

* Remove account_move_chatter

Module account_move_chatter is not relevant in version 14.0.
There is already a chatter on account.move.

* Hide unported modules

* Remove module account_move_reversed_entry

The field reversed_entry_id was added in version 14.0.

* Port account_move_reversal_access to version 14.0

* Port account_move_unique_reversal to v14

* Port account_negative_debit_credit to v14

* Port account_payment_cancel_group to v14

* Move account_payment_from_move_line to unported modules

account_payment_from_move_line would be hard to implement in version 14.0.
account.payment inherits from account.move. Therefore, creating an account.payment from an account.move.line
record would be very tricky

* Port module account_payment_term_usage

* Port account_payment_widget_link to v14

* Port account_report_trial_balance

* Move  account_report_line_menu to unported

* Add account_type_archive (previously account_type_archive

* Port bank_statement_no_reverse

* Move budget_analysis_account_move_line to unported modules

* Port canada_account_types to v14

* Port canada_bank_transfer to v14

* Port payment_list_not_sent

This module is required in version 14.0 for canada_bank_transfer.

* TA#35238 prevent sending twice the same payment

* Begin port hr_expense_tax_adjustment

* Move hr_expense_tax_adjustment to unported modules

Taxes have been complexified a lot in version 14.0.
hr_expense_tax_adjustment was already quite a complex module in version 12.0.
In version 14.0, it would become much more complex and risky.

Therefore, without a major development effort, this module can not be ported to 14.0.

* Port invoice_currency_validation

* Port invoice_fiscal_position_required

* Port invoice_list_email

* Remove invoice_mass_mailing_with_layout

The email is properly rendered in version 14.0.
This module is not anymore relevant.

* Port invoice_refund_not_earlier to v14

* Move invoice_write_access to unported

Will port the module on demand if needed for a given client.

* Port old_accounts

* Remove vendor_invoice_full_list

The module is not relevant anymore in version 14.0. The bug was fixed in Odoo.

* Remove payment_stripe_not_silenced

The code of payment_stripe has improved in version 14.0.
The error seems to be raised properly.

* Update Dockerfile

* Remove module from Dockerfile

* Port invoice_currency_validation to 14.0

* Port bank_statement_partner_name to 14.0

Previously, the module was named bank_statement_extra_columns.
It also added the field note.

This field was renamed narration in v14.0.
The field is already present in the view.

However, it is invisible by default.

* Auto install account_accountant